correctly handle multiple instances of the same variable in midi messages
This affects OSC->MIDI conversions in rules with multiple instances of the
same variable on the right-hand side, such as:

    /1/fader1 f, val : noteon( 0, 127*val, 64*val );

Previously, the value of the variable would only be inserted into one of the
instances on the right-hand side. Now we iterate over *all* instances of the
same variable in the midi message and apply the proper conditioning to each of
them.