use lo_address_new_from_url to set up the OSC return address
This gives more flexibility (use of either UDP or TCP) and delegates the
address parsing to liblo which knows how to do it properly (in particular,
IPv6 addresses should work now). The osc.udp:// prefix is added automatically
if needed, so the previous simplified address format still continues to work.