allow sharing the left-hand side of rules
This allows to abbreviate a collection of rules where the same OSC message is
mapped to different MIDI messages. E.g.,

    /multi/{i} ff, touch,x,y : controlchange( channel, touch, x*127);
    /multi/{i} ff, touch,x,y : controlchange( channel, touch+10, y*127);

may now be written as:

    /multi/{i} ff, touch,x,y : controlchange( channel, touch, x*127);
    : controlchange( channel, touch+10, y*127);

Note that the latter simply gets expanded to the former. There's no change in
semantics (except that an OSC message path can't start with a colon any
longer, but that's rather unlikely anyway).